Retry helper image build
We have seen in the past the `buildx` is sometimes unstable and fails to
build an image because of network connectivity issues or some flaky
resource. In
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/2087 we
added retries for Docker builds.

In https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/2206 we
introduced buildx to build the helper images as well, so we should add
[retries](https://docs.gitlab.com/ee/ci/yaml/#retry) so we don't see
failures like
https://gitlab.com/gitlab-org/gitlab-runner/-/jobs/627685428