Re-enable Docker for windows tests
Some of tests were marked to be skipped in
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/2212 by
mistake. We should still keep running these tests since they run
successfully on Windows and should not ignore them.

Add some of the known failures back to
`ci/.test-failures.servercore1809.txt` even though they are skipped that
file should be our SSOT of which tests we should fix in the future, so
we shouldn't delete the tests even if they are being skipped.

closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/26622