Move GitLab release to its own job
When we want to do a security release, we have to create the tag in a
private fork `gitlab-org/security/gitlab-runner`. When we create the tag
in the private fork we don't want it to create a GitLab release as soon
as the pipeline finishes until the vulnerability is announced publicly.

Move the release create to its own job, have it create the release if the pipeline comes from `gitlab-org/gitlab-runner` and then manual when
pipeline comes from `gitlab-org/security/gitlab-runner`.

Example pipelines for https://gitlab.com/steveazz-playground/security-private/-/blob/ae154fdb7e354c8272d8f52d8fe8813edcc47d54/.gitlab-ci.yml
- Automatically:  https://gitlab.com/steveazz-playground/security-public/-/pipelines?page=1&scope=all&ref=v0.0.6
- Manual: https://gitlab.com/steveazz-playground/security-private/-/pipelines?page=1&scope=all&ref=v0.0.6

Since we are using `rules` for this we have to duplicate `expect_docs`
and `.release` to have a rules equivalent. The naming conventions was
taken from
https://gitlab.com/gitlab-org/gitlab/-/blob/master/.gitlab/ci/rules.gitlab-ci.yml.
Even though we created `.docs:rules:skip` those rules are still being
overridden by the `stable gitlab release` job. I think having this kind
of duplication is better than just updating stable gitlab release to
specify the `before_script` and `stage` which would make it harder when
we start migrating to rules.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/21301#note_384403141