Add security harness
In
https://gitlab.com/gitlab-org/gitlab-runner/-/blob/93156ab755258783ee5059d424887ec0c2705baa/.gitlab/issue_templates/Security%20developer%20workflow.md#L18
we ask the developer to run the security harness script to prevent them
from pushing the public repository.

Copy this script from
[gitlab-org/gitlab](https://gitlab.com/gitlab-org/gitlab/-/blob/master/scripts/security-harness)
to the sake of consistency. This uses a ruby script and ruby is
something we don't use in this project. However it's better to just do
copy paste rather then have our own go/bash version. This script is only
going to be used by team members at GitLab and not by the wider
community since security releases are done by GitLab team members, so
expecting them to have ruby installed is normal.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/21301