Add support for custom PUT HTTP headers in cache archiver
This commit adds a new method in the cache adapter to allow for custom,
unsigned HTTP headers. These headers are passed via the --header
parameter to the cache archiver process. The default behavior of adding
Content-Type and Last-Modified headers is retained if the --header
parameters is omitted.

This is in preparation for supporting the Azure PUT Blob API
(https://docs.microsoft.com/en-us/rest/api/storageservices/put-blob),
which needs the `x-ms-blob-type` HTTP header, which cannot be sent via a
query string.