Add variable to enable fallback cache key
This change introduces a CACHE_FALLBACK_KEY variable that
is used for a second attempt when the cache fails to download
with the default key. By e.g. setting the fallback key to
'master', this provides a way for each branch or merge request to
use its own cache key, while completely avoiding the lengthy first
build, since the cache is simply cloned from master if it does
not yet exist.