git: Override HTTP user agent
For analytics it's nice to understand who is interacting with Git
repositories. For example in the case of the GitLab Runner, it uses the
same interface any other user has as it uses the normal Git. This makes
the metrics gathering harder. For example, for GitLab.com we currently
use Google IP ranges as heuristic.

This change overrides the git user agent, so for SRE's and Gitaly
developers it's easier to understand what clones are from GitLab-CI,
versus the other clones. It includes the runner version most notably, as
well as further runner details, as OS and Go version.