Doc `cleanup_file_variables` for custom executor
In https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27001 we had to
introduce a new step for the runner to execute at the end of the build
to clean up the file based variables from disk. We need to document this
step on the custom executor so that users are aware of it.