Skip CleanupFileVariables stage if no file variables
In https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27001 we
introduced a step that will clean up file based variables. However if we
are still running this stage even if there is no file based variables
defined. If we keep running the step no matter what we are wasting
compute resources and it wouldn't be consistent with our other steps
where they are skipped it it would be a no-op.

Update the test to create a brand new `build` and `info` so the
JobVariables aren't "cached" from the previous run.

closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27057