Use e.helperImageInfo to import helper image
The docker executor will first look at known paths on disk to see if it
can import the helper image. We are using a hardcoded name for when we
import the image. The name can now change depending on
`FF_GITLAB_REGISTRY_HELPER_IMAGE`. Use the executor helperImageInfo to
import the image with the correct name since that is the single source
of truth for the helper image.

Add full coverage of the `getLocalHelperImage` method and the methods it
is calling since we don't have any coverage for that.