Fix check_modules command
The check_modules should check if there are any changes not committed in
`go.sum` or in the `vendor` directory. However, this requires to have no
changes in `go.sum`, if there are changes this wouldn't work because we
compare the diff before the command and after the command.

The variable [`OUR_PACKAGES` is running `go
list`](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/11e965397800fa0a3cd457953fc7e18568651e95/Makefile#L40)
which runs `go mod tidy` behind the scenes. This results into a dirty
`go.sum` and when we run `check_modules` the changes are already present
so we don't see any new changes.

Remove the usage of `OUR_PACKAGES` since it's no longer needed, we don't
have to filter out the `vendor` directory anymore and using `./...` is
enough in newer versions of Go.

This will also make our Makefile faster, because we don't have to run
`go list` for every command.

Make sure we reset the `go.sum` and the `vendor/` dir because we run the
git diff command to prevent any failures like these in the future.

closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27229