Fix Azure cache not working in K8S executor
In the K8S executor, the object name is prefaced with the cache
directory (e.g. `/cache`), which is then used to sign the object
name. However, the cache archiver discards leading slashes when
uploading the file, so Azure rejects the upload with a signature
mismatch error.

To fix this, we normalize object names by dropping leading slashes.

Closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27324