Add object storage cache credentials adapter
This is the first merge request split from
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/2639 to
provide static credentials for S3 via environment variables to the
cache helper.

Note that this change does not yet export these variables. It is
intended to lay the groundwork for using Go Cloud to upload cache files
via the AWS SDK client.