Improve logging to packagecloud push
When debugging
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27470 it was hard
to understand what is going on inside of the job log like
https://gitlab.com/gitlab-org/gitlab-runner/-/jobs/974893081 it's not
really clear what we are pushing for which distro.

Update to log the distro that we are going to push and also enable
verbose mode to get more detail of what we are pushing.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27470