fix(test): version pin pwsh version
The `alpine-3.12` tag points to the latest version of PowerShell core, this
gets automatically updated when a new version is released, for example
in
https://gitlab.com/hirokiyoshida837/gitlab-runner/-/jobs/1025765632#L2066
it failed because a new version of PowerShell (7.1.2) came out and we are
expecting v7.1.1.

Use the specific PowerShell version so we make sure our tests run
successful for that version and aren't subject to updates.

closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27554