Add integration tests for job trace limit
The output limit exceeding should not fail the job. When the limit is
exceeded Runner prints the message that it happened and stops collecting
output, but that's all. Job keeps executing and the final
failure/success should be detected as usual.

A recent change have introduced a regression that causes jobs that were
properly passing with job output limit exceeded warning on 13.8.0
to fail on 13.9.0-rc1.

This commit is adding integration tests that are covering these cases:

1. Job with exceeded output limit that normally should success: the
   warning message is printed, execution of job after printing the
   message is happening, the job is succeeded.

2. Job with exceeded output limit that normally should fail: the warning
   message is printed, execution of job after printing the message is
   happening, the job is failed.

3. Job with exceeded output limit that is being canceled: the warning
   message is printed, execution of job after printing the message is
   happening until cancel is requested, the job is canceled.