Remove unused context usage
Context passed to `ExecOptions.executeRequest()` seems to be added
to allow cancelation of executed request.

Unfortunately, the context is used only by adding it to the request
creator. But the request object is used by Runner only to create
the URL. We don't call the `req.Do()` which would execute the request, we
just call `req.URL()` and pass this URL to the `executor.Execute()`
method. This is the method that really runs the request and where
cancelation would need to be handled.

Context passed from the method call to the request, is not used here at
all.

Unfortunately, the `remotecommand.Stream()` that is the finall call
in `executor.Execute()` doesn't have any canceling mechanism nor is able
to take the usage of the context.

Therefore the context here is totally unneeded and should be removed.

If we want to make it possible to cancel these requests, we need to
start from figuring out how to properly cancel `remotecommand.Stream()`.