Remove docker-machine provision on creation failure
In
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27613#note_522938477
we see machines failing to create on the cloud provider, but it fails
for rate limit reasons or anything else. This leads to a split-brain issue where we have state of a machine "created" in `docker-machine`
state but the machine is nonexistent/inaccessible to the cloud provider.

If the creation fails `docker-machine provision` runs on a machine that
doesn't exist which wastes API requests, and time trying to wait for SSH
to be available on a machine. The `docker-machine provision` tries to
re-provision an existing machine, but if creation fails that means that
the machine doesn't exist.

Instead of retrying to create/provision the machine with the same details
fail fast, remove the machine and try to create a new one with details.
This should result in much quicker recovery when we reach the API rate
limits on the cloud provider and also job queues because the machine
would be available sooner rather than waiting for `docker-machine
provision` to be available.

To enable this fail fast functionality you have to enable
`FF_SKIP_DOCKER_MACHINE_PROVISION_ON_CREATION_FAILURE` feature flag
inside of the `config.toml`

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27613