Don't return error when checking feature flag
When the value of the feature flag can't be parsed an error is return.
This error is being ignored for most cases because there isn't any value
in handling that error.

The only time the error is handled is for logging. Instead of logging it
in 1 place move the log inside of the `featureflag` pkg where the caller
passes the logger.

With this we get the added benefit of using the logger that is already
initialized, not that there was anything preventing us from doing so in
the past. For example when we fail to parse the value of a specific job

Before:
```
ERROR: Error while parsing the value of feature flag  error=strconv.ParseBool: parsing "asdf": invalid syntax name=FF_RESET_HELPER_IMAGE_ENTRYPOINT value=asdf
```

After:
```
ERROR: Error while parsing the value of feature flag  error=strconv.ParseBool: parsing "asdf": invalid syntax job=2187 name=FF_RESET_HELPER_IMAGE_ENTRYPOINT project=20 runner=4B72igtx value=asdf
```