feat(ff): set feature flag in config.toml
What:
Allow user to set a feature flag inside of the config.toml with
`[runners.feature_flags]` which aren't overridable able by an job
variables.

Why:
In https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/2805 we
want to change how `docker+machine` executor creates machines and we
want to roll this out behind a feature flag. At the moment users can
only specify a feature flag as a job variable, which won't work in this
case because a machine is created before a job exists so there is no way
we can set this.

part of https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27613