Complete the example configuration for gcp cache
The example given in the documentation does not mention that the user
has to add a name entry for `secrets` that is needed for the example to
work.
By adding the entry, the JSON file will be successfully mounted in the
runner at `/secrets/gcs-application-credentials-file`
When this entry isn't present, the runner won't mount the JSON file with
GCP credentials, and the user will only see an error:
"No URL provided, cache will not be downloaded from shared cache server.
Instead a local version of cache will be extracted."

Changelog: changed