refactor(ci): remove unset GPG_KEY
With https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27963 the
variables were moved to be file-based so there is no longer a need to
unset this variable. There is also a plan to scope these variables for
specific jobs through environments.

There was already a problem with this unset from a maintainability
perspective because not all jobs had it or it wasn't safe enough for us.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27963