ci: update env name for linux docker images
Part of https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27963 we
are
[scoping](https://docs.gitlab.com/ee/ci/variables/#limit-the-environment-scope-of-a-cicd-variable)
certain environment variables to jobs so they are only available in that
job. For `DOCKER_HUB_PASSWORD` and `DOCKER_HUB_USER` we already scoped
it for `stable/docker_images/*` in
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27963#note_601887655
but this doesn't match the `stable/docker_image` environment for the
linux containers.

Update the name to match the wildcard scope.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27963