refactor: only print necessary env vars in tests
When using the `env` or `set` keyword it prints all the environments
that are available in the job. These can be problematic printing
environments that we shouldn't print. Update the tests to run an `echo`
on the specific variables that we are testing instead. This makes
running the tests safer since it won't print in every environment.

In https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/2951 we
only allow certain environment variables, however, it's better to have
multiple layers of protection rather than 1 just in case 1 of them
regresses.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27963