fix: make prepare_index read GPG_KEY from file
Background
---
In https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/2958 we
changed the `GPG_KEY` and `GPG_PASSPHRASE` to be file based CI/CD
variables rather then environment base. However we started seeing
failures that the gpg key could not be loaded such as
https://gitlab.com/gitlab-org/gitlab-runner/-/jobs/1372115967.

In
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/28008#note_610274443
we see that when we use `cat` it doesn't escpae the newlines properly.

Fix
---
We can use the
[`file`](https://www.gnu.org/software/make/manual/html_node/File-Function.html)
function to fix this, but then we can't have condtional logic where we
only define `GPG_KEY` when the key exists like we want too, since we
don't want to sign development packages.

Follow the same logic we have in
[package.gitlab-ci.yml](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/327a951d1d52a414b07c7c0d44976cd494f71b9f/.gitlab/ci/package.gitlab-ci.yml#L11-19)
to only add it if it's defined.

reference https://gitlab.com/gitlab-org/gitlab-runner/-/issues/28008