Migrate to github.com/creack/pty and bump used version
github.com/kr/pty was a fork of github.com/creack/pty. The fork
is discontinued and already archived by the owner. Development
keeps happening in the upstream repository.

Recent versions have some updated. One of them fixes a problem
- related to TTY usage - that we've seen in our tests after migrating
to Go 1.16.x.

With this commit we're switching to the main fork of the library
and updating its version from 1.1.5 to 1.1.11.