Automatically retry integration_k8s jobs
The job have known flaky tests that we haven't yet fixed. It's already
annoying when it fails the pipeline (especially the merge train ones!)
because of that. And given our current experiment on Continuous Delivery
of GitLab Runner to our maintained fleet
(https://gitlab.com/gitlab-org/gitlab-runner/-/issues/28713) it becomes
even more annoying.

We already use `retry: 2` in few of our jobs that are known from failing
randomly on unknown reasons. This commit follows the same approach.

Merging this doesn't of course mean that we shold not identify and fix
what's causing `integration_k8s` random failures. It just aims to make
the job less annoying for the maintainers until this is done.