Move the "static QA" job to the postrelease stage
The `static QA` job was added a long time ago to make sure,
that the pipeline with linting violations is failing.

We didn't want however tese violations to prevent the change from being
fully tested, compiled and released - especially in the development
version where the packages/binaries are still useful for testing even
when the code requires quality updates.

Therefore we've marked the `code_quality` job as allowed to fail and
added this `static QA` one that fails when the code quality analysis
artifact contains any information. The job was added to - at that time -
the last stage in the pipeline.

Since then however we've introduced the `postrelease` stage, where this
job fits definitely better. And this patch is doing exactly that - it
moves the `static QA` to the postrelease job. It still may fail the
pipeline (which should prevent use from merging changes decreasing code
quality) while it will allow most of the pipeline to run without
interruption.