Show error details for failed artifact uploads
When the runner receives a 400 Bad Request when uploading CI artifacts,
the user does not receive any feedback as to what was wrong with the
upload.

We can fix this by parsing the API JSON response. If a `message` exists,
show that since this is intended to be a human-readable message.

Relates to https://gitlab.com/gitlab-org/gitlab-runner/-/issues/28743