Document need for entrypoint to open shell
In 712da21f53162599e285ce3c43228d9a7a6106ea, the Kubernetes executor
will start to honor the entrypoint that is defined in the docker image.
However, one requirement is that it needs to open a shell in order for
the tests to run. This change documents that.

Changelog: added