Divide packages buildling jobs in the pipeline even more
With the size of the package increased significantly over
last months (mostly because we now bundle a number of dumped
container images for the helper container), current matrix
of package builds exceeded the 1GB limit of upload on GitLab.com.

This becomes a blocker for community contributors who run
pipelines in their own projects.

This commit refactors the matrix a little to make sure that
none job exceeded 1GB size of the artifacts to upload.