Align debian releases for stable and Bleeding Edge versions
With https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/2940
we've replaced `debian/jessie` with `debian/bullseye` release for the
DEB package. This however happened only for the list defined for
Bleeding Edge version of the Runner and not in the list that defines
DEB releases for Runner's stable version.

This commit aligns these two lists and adds missing update in the
main list maintained in the Makefile.