Support Vault EE namespaces
Currently the vault integration does not support Vault EE namespaces.
Only the vault server URL, role name and auth path can be configured
with the variables. It is possible to just prefix the engine's path
with the namespace, but it is not possible to specify the namespace
for the authentication.

This change makes the namespace configurable wiht the VAULT_NAMESPACE
variable (similar to how `VAULT_SERVER_URL`, `VAULT_AUTH_ROLE` and
`VAULT_AUTH_PATH are being used already). API operations performed
under a namespace are done by providing the namespace name in
`X-Vault-Namespace` header.