Log object storage forbidden errors during artifact downloads
When artifact downloads are redirected to an object storage provider,
the runner can encounter a 403 Forbidden error. It can be difficult to
understand whether this 403 error was due to a permissions problem on
GitLab or an issue with object storage.

We had one customer that had enabled Google's VPC Service Control, and
used it to blocked downloads from other buckets. However, it took us a
while to figure out the 403 was caused by this feature.

When object storage responds with an error to a GET request, most
providers (Amazon, Google, and Azure) return a standard XML message
that contains both a code and a message. We now attempt to decode a
JSON or XML response when a CI artifact download encounters a 403
error and display the message in the output.

Changelog: added