[common] Method to determine pull_policy source
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/29115 asks for the
error message for incompatible pull_policy and allowed_pull_policies to
indicate the source of the pull_policy configuration. This method will
help with that by computing the source of the pull_policy configuration,
one of: the project's giltab-cy.yaml, the runner's config.toml or the
runner's default config value.

The best we can do is to indicate if the used config value came from the
project's or the runner's configuration. We can't for e.g. identify the
configuration section within a file, nor can we narrow down an
individual file when the project's config file imports configurations
from other files.