[common] Error type for failure to validate pull policies
We'll use this to nicely format the error message, which can take on
different forms depending on the source parameter.

Note that the `pullPolicy` and `allowedPullPolicies` struct members and
arguments to the constructor are currently `interface{}`, with a note to
use generics to make this better when we move to Go 1.18.x.

We need generics (or empty interface) because the types of those two
variables are different in the `docker` and `kubernetes` packages; in
the former the type is `common.DockerPullPolicy` (or
`[]common.DockerPullPolicy`) whereas in the later it is
`k8s.io/api/core/v1.PullPolicy`. However, both types are just aliases of
`string`, so they are totally compatible.

In any case, the only use of the variables is to format them for
printing via `%v`, so an empty interface actually works fine here.
Judicious use of generics would be nicer though.