Update linter version to 1.46.2
We'll have to do this to update to go 1.18, and we have to do it now to
match the linter version in the `goargs` repo. The versions of `goargs`
and the linter have to match because the latter loads the former as a
plugin, and plugins dictate that the version of common dependencies
across the a plugin and its consumer must be the same.