Revert github.com/urfave/cli back to v1.20.0
As pointed at https://gitlab.com/gitlab-org/gitlab-runner/-/issues/29193#note_1039039254,
update of github.com/urfave/cli to v1.22.2 was probably unwanted
and made by mistake. At the same time, it brought a regression, as
the CLI flags parser works now slightly differently than it was.

If we will decide we want to update this dependency, we should make
some tests and check if it changes any other behavior. And we should
adapt our usage of it accordingly.

For now, this commit will revert the update, which will resolve the
reported regression.