[docker] Method to capture logs for single service container
Given a container's ID, this method will hook into the container's log
stream and copy all logs to the specified sink, for the duration of the
container's life. `stdcopy.StdCopy` handles the actual reading from the
container and writing to the sink, and will block until either the
connection to the container is closed or a read or write error occur.

Note that this approach uses the docker SDK's `ContainerLogs()` API; we
could have also used the `ContainerAttach()` API with basically the same
code here. Both APIs return an `io.Reader` which can be copied via
`StdCopy()`. I opted for `ContainerLogs()` because it's a slightly
simpler API.

Since `StdCopy()` blocks for the life of the service container, we have
to run that call in a goroutine.