[docker] Method to set up copying service container logs
This method will, for each configured (helper) service, call
`captureContainerLogs()` to set up log capture. In this case, logs are
redirected to the current task's trace log (i.e. `e.Trace`) via an
`InlineServiceLogWriter` instance (to prefix logs with the container's
name).

In the future, if we want to redirect the logs elsewhere (e.g. a file to
be subsequently captured as a CI artifact), we just have to change the
`sink` passed to `captureContainerLogs()`.