[kubernetes] Method to capture logs for single service container
Given the name of a kubernetes-managed container (and a reference to the
pod), this method will hook into the container's log stream and copy all
logs to the specified sink, for the duration of the container's life.

This is analogous (and as similar as possible) to what is being done in
the docker executor to capture service logs there.