[kubernetes] Method to set up copying service container logs
This method will, for each configured (helper) service, call
`captureContainerLogs()` to set up log capture. In this case, logs are
redirected to the current task's trace log (i.e. `e.Trace`) via an
`InlineServiceLogWriter` instance (to prefix logs with the container's
name).

Again, this is analogous (and as similar as possible) to what is being
done in the docker executor to capture service logs there.