[kubernetes] Service log capture integration tests
Three tests: one to make sure the feature is not enabled when
CI_DEBUG_SERVICES is not set, one to make sure the variable is handled
correctly, one to make sure the feature works (i.e. service container
logs are captured and written to the main trace log, with prefix).