Allow setting of Docker volume label mode independent of read/write mode
Previously attempting to mount a volume with "ro,z" or "rw,z" would
fail because the Linux mount parser only allowed one option.  It is
sometimes necessary to use both the `ro` and `z` flags at once when
mounting a volume in a container, such as when the host has SELinux
policies.

Relates to https://gitlab.com/gitlab-org/gitlab-runner/-/issues/29247