Delete tmp/gitlab-test directory.
Runner tests depend on the gitlab-test repository being available and
the `development_setup` target checks for the directory's
existance. However an empty gitlab-test repo was checked in so
development_setup wasn't cloning the repo, causing some tests to fail.

This change removes the empty repo and adds cleanup of the repo to the
`clean` target.