[bash] Refactor BashWriter.TmpFile()
Move everything but the path concatenation to a new method, which we'll
use later.

Why do we need to do this at all? That tmp file path set by/in
JobVariables is an approximate path. TO actually use it in the shell
writer we need to tweak it a little. We'll use cleanPath() for that.