[common] Correctly handle file variables
The notion of file variables is arguably incomplete in the JobVariable
implementation. There are things this type of object needs to be able to
do (like generate a temporary file path), which it previously could not,
which results in behaviour like
https://gitlab.com/gitlab-org/gitlab/-/issues/29407. This is the first
step in fixing this.

The _value_ of the RUNNER_TEMP_PROJECT_DIR variable is some other
object's responsibility, but once having been injected into a
JobVariables instance, it's the latter's responsibility to create
temporary file paths.

Contrary to the suggestion in
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/29128, This strikes
me is a more strictly correct place for this method than the Build
object since the JobVariables object shuld be responsible for knowing
which of its fields (Key, Value, ???) should be used to generate the
temporary file path, while the Build object is responsible for
specifying the builds temporary path.

Now that the JobVariables object has a way to generate temporary file
paths, we can correctly handle expanding both file variables and
variables that reference (or point to) file variables. The effect of
this change is that when we call `Get()` of a file variable, we'll get
back the temp file path instead of the (intended) content of the file.
This automatically fixes variables that reference file variables, since
expansion of that type of variable will result in the file variable's
tmp file path instead of its content. 

This mostly fixes https://gitlab.com/gitlab-org/gitlab/-/issues/29407,
making some minor tweaks in the ShellWriters (later).