[common] Inject tmp project dir into JobVariables
This is a convenient place to put the tmp project dir such  the it can
be easily accessed by methods on JobVariables. We'll use it next to
correctly expand variables that reference job variables.

After a few attempts, this suggestion by @ajwalker is the cleanest way
to inject the build tmp path and tmp file creation into the JobVariable
or JobVariables object.