Add go:generate directives to generate mocks
This is the minimum set of mocks required (i.e. actually used by tests).

I've added the directive directly above the interface that needs to be
mocked. Another alternative is to put the directive in the test file
that uses the mock, but there are cases where a mock is used in more
than one file, and other where the mock is used outside of the package
in which the corresponding interface is defined, both of which made
things more awkward than the approach taken here.