Simplify installation of mockery
Use `go install` to install mockery instead of downloading the tarball
(for the right or/arch), extracting it, and making it executable. `go
install` can do all that for us, so we may as well take advantage of it.

Note that to handle the upgrade case (i.e. ensure the new version is
installed when we change the version), I went with the dead simple
approach of _alwasy_ installing mockery when this target is run.
Generating mocks happens infrequently enough, and installing mockery via
`go install` is fast enough that this approach is fine.