[helper] New inline service `io.Writer` type
This simple `io.Writer` prefixes log messages/lines with the service
container's name, and adds some colour to the logs. We'll use it later
on to stream service container logs into the tasks main trace log.

The prefix and colour will help to differentiate the service container's
logs from the main task's trace log, and from other service container
logs.

In truth the name of this type could be a little more generic, since
it's actually a property of the supplied sink that makes the final
`io.Writer` an "inline" writer (whatever that means). But int he context
of these changes, this writer is intended to (and will be) used to write
service container logs inline into the main trace logs, hence the name.