[helper] Testing util
I'll use this in the next commit to test the `InlineServiceLogWriter`
type introduced in the previous commit. Specifically, in the success
case the number of bytes written returned by `Write()` must be exactly
the length of the message, and if not all bytes were written, `Write()`
must return an error.