Specify full paths of go-fips image in BUILD_IMAGE var
If we only specify the image name (w/o the image registry path), the
security scan that runs agains this image in
https://gitlab.com/gitlab-com/gl-security/appsec/container-scanners will
fail because it tries to `docker pull` `go-fips:<tag>`.

Yes, this here is only a default value, and the correct full name is
specified elsewhere, but may as well fix the default value if we know
it's broken.