Ensure the go-fips image is rebuilt when we upgrade Go
More specifically, ensure a build of the `go-fips` image is triggered
when the `GO_FIPS_VERSION` variable is changed.

The rules to rebuild the `go-fips` image are when any of the following
files are changed:
```
- dockerfiles/ci/go.fips.Dockerfile
- dockerfiles/ci/go.fips.Dockerfile.rebuild
- .gitlab/ci/prepare.gitlab-ci.yml
```

However, the `GO_FIPS_VERSION` variable is defined in
`.gitlab/ci/_common.gitlab-ci.yml`, which is imported and used in
`.gitlab/ci/prepare.gitlab-ci.yml`. Because `_common.gitlab-ci.yml` is
not in the above list, changing the value of that variable won't trigger
a rebuild of the image, which is wrong.

So, add `.gitlab/ci/_common.gitlab-ci.yml` to the build-rules for the
`go-fips` to ensure it's built when `GO_FIPS_VERSION` changes.

This should fix 29369.