Skip docker Test_CaptureServiceLogs integration tests on windows
Making this test run on windows is not feasible because most of the
windows environments (1809, 2004, 20H1, 20H2) require a slightly
different version of whatever service is used for the test (to match the
OS version/license).

Ironically though, the logs from some of the (failed, for other reasons)
test runs show the feature does work on windows.