Prevent taskscaler thrashing instances
The change first uses taskscaler.Reserve() to reserve capacity, and once we have
a job, calls taskscaler.Acquire() in the wrapped executor returned by the
executor provider. This ensures that when idle scaling is enabled, we only
accept jobs once we've confirmed there's capacity and have reserved it.

Previously, we were acquiring an capacity and then releasing it if there was no
job. But this thrashed VMs if taskscaler had been configured to remove capacity
after it has been used.