Script to print OS distribution/versions for which we want to publish gitlab-runner  packages
This (go) script will query the packagecloud API for a list of all the
distro/releases it supports, and whitle that list down to:
1. the package type (deb or rpm) specified in the command invocation.
2. the branch (stale or bleeding) specified in the command invocation.
3. the distros for which we claim to provide runner packages as encoded in
   supportedDistrosByPackageAndBranch.
4. releases of the distro we want to support (not too old, preferably LST) as
   encoded in oldestRelease and skipReleases.

Finally the script will print that result with no decorations.

This output will, in the next commit, replace currently hard-coded and
manually updated RPM_PLATFORMS and DEB_PLATFORMS variables.

I've intentionally kept the number of 3rd party deps here to 0 to make
sure it builds quickly, and so we don't need a go.mod

Run it like so:

PACKAGECLOUD_TOKEN="..." PACKAGE_CLOUD_URL="..." go run ./scripts/packagecloud-releases.go deb stable