Automate packagecloud distro releases
With this change, the distro releases for which we publish runner
packages (deb ad rpm) will no nonger be hard coded in RPM_PLATFORMS and
DEB_PLATFORMS (and require manual updating to keep up with
packagecloud). Instead, the distro releases will be determined by
`scripts/packagecloud-releases/main.go` and be based on whatever
releases our packagecloud instance currently supports.

This means no manual intervention to:
- add new releases
- remove old/EOL releases
- keep documentation up to date

BTW, all of the above are currently out of sync with packagecloud and
each other.

What we loose in exchange for the automation is:
- control over exactly when EOL releases stop being supported. The
  current value of NUM_DISTRO_RELEASES is more eager to drop old
  releases than our prevous manual approach (no surprise). But we were
  claiming to support already EOLed releases, so this is fine and good.
- Explicit documentation of exactly what releases are supported. This was
  also out of sync with reality, so this is acceptable if not ideal. We
  could, if we choose to, update the documentation manually at each
  release based on what the script outputs.
- The ability to selectively skip distro/release if for whatever reason
  we won't want to support one.

I'd argue this is a worthwhile tradeoff for keeping up to date with
packagecloud.