Fix computation of test unit count
There's arguably a bug in the way that powershell's `[int]` operator
converts floats to ints. Specifically, `[int](4.5) == 4` but
`[int]4.5000000001` == 5. I'd expect (reasonably I think) for int to
have either `floor` or _proper_ `round` semantics.

Anyway, the result of this bug was that some windows integration tests
(possibly other tests too) were being silenty no run. This is because
there are currently 27 integration tests, and with the fixed 6 test
units (i.e. `CI_NODE_TOTAL`), `(27/6) == 4.5`, but `[int]4.5 == 4`. So
the last three integration tests were being silently dropped.

This change fixes the above problem, but it turns out that some of the
tests that were being dropped fail reliably! I'll address this in the
next commit by adding to the list of expected failures.